Fix bug in serviceAccount DiffSuppressFunc
The DiffSuppressFunc on the service_account field of
google_compute_instance was operating on all sub-elements in the list,
not just on the number of elements in the list, like it meant to. This
caused panics when trying to type assert old and new values. This
updates to only run the function when comparing the number of elements
in the list.